Refactor jedis instrumentations
