--- dd-java-agent/instrumentation/jedis-1.4/src/main/java/datadog/trace/instrumentation/jedis/JedisClientDecorator.java
@@ -1,44 +1,0 @@
-package datadog.trace.instrumentation.jedis;
-
-import datadog.trace.agent.decorator.DatabaseClientDecorator;
-import datadog.trace.api.DDSpanTypes;
-import redis.clients.jedis.Protocol;
-
-public class JedisClientDecorator extends DatabaseClientDecorator<Protocol.Command> {
-  public static final JedisClientDecorator DECORATE = new JedisClientDecorator();
-
-  @Override
-  protected String[] instrumentationNames() {
-    return new String[] {"jedis", "redis"};
-  }
-
-  @Override
-  protected String service() {
-    return "redis";
-  }
-
-  @Override
-  protected String component() {
-    return "redis";
-  }
-
-  @Override
-  protected String spanType() {
-    return DDSpanTypes.REDIS;
-  }
-
-  @Override
-  protected String dbType() {
-    return "redis";
-  }
-
-  @Override
-  protected String dbUser(final Protocol.Command session) {
-    return null;
-  }
-
-  @Override
-  protected String dbInstance(final Protocol.Command session) {
-    return null;
-  }
-}
