--- dd-java-agent/instrumentation/jedis-3.0/src/main/java/datadog/trace/instrumentation/jedis3/JedisClientDecorator.java
@@ -1,45 +1,0 @@
-// Modified by SignalFx
-package datadog.trace.instrumentation.jedis3;
-
-import datadog.trace.agent.decorator.DatabaseClientDecorator;
-import datadog.trace.api.DDSpanTypes;
-import redis.clients.jedis.commands.ProtocolCommand;
-
-public class JedisClientDecorator extends DatabaseClientDecorator<ProtocolCommand> {
-  public static final JedisClientDecorator DECORATE = new JedisClientDecorator();
-
-  @Override
-  protected String[] instrumentationNames() {
-    return new String[] {"jedis", "redis"};
-  }
-
-  @Override
-  protected String service() {
-    return "redis";
-  }
-
-  @Override
-  protected String component() {
-    return "redis";
-  }
-
-  @Override
-  protected String spanType() {
-    return DDSpanTypes.REDIS;
-  }
-
-  @Override
-  protected String dbType() {
-    return "redis";
-  }
-
-  @Override
-  protected String dbUser(final ProtocolCommand session) {
-    return null;
-  }
-
-  @Override
-  protected String dbInstance(final ProtocolCommand session) {
-    return null;
-  }
-}
