+++ dd-java-agent/instrumentation/jedis/jedis-3.0/src/main/java/datadog/trace/instrumentation/jedis3/Jedis3Instrumentation.java
@@ -1,7 +1,7 @@
 // Modified by SignalFx
-package datadog.trace.instrumentation.jedis;
+package datadog.trace.instrumentation.jedis3;
 
-import static datadog.trace.instrumentation.jedis.JedisClientDecorator.DECORATE;
+import static datadog.trace.instrumentation.jedis3.Jedis3ClientDecorator.DECORATE;
 import static java.util.Collections.singletonMap;
 import static net.bytebuddy.matcher.ElementMatchers.isMethod;
 import static net.bytebuddy.matcher.ElementMatchers.isPublic;
@@ -10,24 +10,23 @@ import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
 
 import com.google.auto.service.AutoService;
 import datadog.trace.agent.tooling.Instrumenter;
-import datadog.trace.api.Config;
 import io.opentracing.Scope;
 import io.opentracing.util.GlobalTracer;
-import java.io.UnsupportedEncodingException;
 import java.util.Map;
 import net.bytebuddy.asm.Advice;
 import net.bytebuddy.description.method.MethodDescription;
 import net.bytebuddy.description.type.TypeDescription;
 import net.bytebuddy.matcher.ElementMatcher;
 import redis.clients.jedis.Protocol.Command;
+import redis.clients.jedis.commands.ProtocolCommand;
 
 @AutoService(Instrumenter.class)
-public final class JedisInstrumentation extends Instrumenter.Default {
+public final class Jedis3Instrumentation extends Instrumenter.Default {
 
   private static final String SERVICE_NAME = "redis";
   private static final String COMPONENT_NAME = SERVICE_NAME + "-command";
 
-  public JedisInstrumentation() {
+  public Jedis3Instrumentation() {
     super("jedis", "redis");
   }
 
@@ -42,7 +41,8 @@ public final class JedisInstrumentation extends Instrumenter.Default {
       "datadog.trace.agent.decorator.BaseDecorator",
       "datadog.trace.agent.decorator.ClientDecorator",
       "datadog.trace.agent.decorator.DatabaseClientDecorator",
-      packageName + ".JedisClientDecorator",
+      "datadog.trace.instrumentation.jedis.JedisClientDecorator",
+      packageName + ".Jedis3ClientDecorator",
     };
   }
 
@@ -52,7 +52,7 @@ public final class JedisInstrumentation extends Instrumenter.Default {
         isMethod()
             .and(isPublic())
             .and(named("sendCommand"))
-            .and(takesArgument(1, named("redis.clients.jedis.Protocol$Command"))),
+            .and(takesArgument(1, named("redis.clients.jedis.commands.ProtocolCommand"))),
         JedisAdvice.class.getName());
     // FIXME: This instrumentation only incorporates sending the command, not processing the result.
   }
@@ -61,22 +61,15 @@ public final class JedisInstrumentation extends Instrumenter.Default {
 
     @Advice.OnMethodEnter(suppress = Throwable.class)
     public static Scope startSpan(
-        @Advice.Argument(1) final Command command, @Advice.Argument(2) final byte[][] args) {
-      final Scope scope = GlobalTracer.get().buildSpan("redis." + command.name()).startActive(true);
-      DECORATE.afterStart(scope.span());
-      String statement = command.name();
-      if (args.length > 0
-          && !statement.toLowerCase().equals("auth")
-          && Config.get().isRedisCaptureCommandArguments()) {
-        statement += ":";
-        for (final byte[] word : args) {
-          try {
-            statement += " " + new String(word, "UTF-8");
-          } catch (UnsupportedEncodingException e) {
-          }
-        }
+        @Advice.Argument(1) final ProtocolCommand command,
+        @Advice.Argument(2) final byte[][] args) {
+      String commandName = "query";
+      if (command instanceof Command) {
+        commandName = ((Command) command).name();
       }
-      DECORATE.onStatement(scope.span(), statement);
+      final Scope scope = GlobalTracer.get().buildSpan("redis." + commandName).startActive(true);
+      DECORATE.afterStart(scope.span());
+      DECORATE.onStatement(scope.span(), commandName, args);
       return scope;
     }
