+++ settings.gradle
@@ -59,8 +59,9 @@ include ':dd-java-agent:instrumentation:java-concurrent:akka-testing'
 include ':dd-java-agent:instrumentation:java-concurrent:akka-2.5-testing'
 include ':dd-java-agent:instrumentation:jboss-classloading'
 include ':dd-java-agent:instrumentation:jdbc'
-include ':dd-java-agent:instrumentation:jedis-1.4'
-include ':dd-java-agent:instrumentation:jedis-3.0'
+include ':dd-java-agent:instrumentation:jedis'
+include ':dd-java-agent:instrumentation:jedis:jedis-1.4'
+include ':dd-java-agent:instrumentation:jedis:jedis-3.0'
 include ':dd-java-agent:instrumentation:jersey'
 include ':dd-java-agent:instrumentation:jetty-6'
 include ':dd-java-agent:instrumentation:jetty-8'
